Add advanced-git-search.nvim
This plugin is added to do extended search regarding the git history of a repository.
It allows us to search the history of the entire file or just the selected lines.
It can also show us the diff between the history and current file.

It can be thought of the final piece of the git integration along with lazygit and gitsigns.

The keybindings are not added yet, they will be added in a future commit.