Add placeholders folder to contain templates for certain plugins
The folder is currently added for VRC only since it requires a .rest file to work.

The placeholder file is really bare bones and only showcases a single request.
If desired, this file can be extended to cover all features of VRC.

Since I use Postman in a really basic way, I wanted to use VRC just like that as well.

The design of this .rest file is all up to the users, you can have a .rest file for each of your endpoint, or have one .rest file which contains all.