Add a command to generate rest template for .rest files
A new function is written for VRC to automatically generate .rest template.

This function takes the template from .config/placeholder file and copies it to the current nvim buffer.

Later on, this functionality will be expanded for other file types. Dynamic template searching with fzf will be added as well.