[setup] Init dotfiles setup
A setup directory is created to house necessary Ansible playbooks and
tests for the setup (via Vagrant).

Also, a Makefile is added to simplify the setup and test scripts.

As a start, a pre_tasks task is added to main playbook, which configures
the package managers for both MacOS and Ubuntu.