Make .zshrc support cross platforms
The PATH is updated by spliting Homebrew lookup into another line.
This will help with non-Darwin distribution setups.

Also, brew prefix is removed from zsh plugins to use this .zshrc file
for non-Darwin distributions.

Last but not least, starship initialization is fixed, if the binary is
already initialized, then it won't be initialized again. This is done to
fix the FUNCNEST issue when using vim mode in shell.