[setup] Install Hashicorp tools for Darwin.
A new role called acikgozb.hashicorp is added to install Hashicorp
related tools to the host.

Since all Hashicorp tools are installed more or less in the same way, it
made sense to create a new role for it.

Currently, only Darwin implementation is done, it was the more complex
one due to having both .dmg and .zip installations.

Linux installations will be done in a later commit.