[fedora] Replace Docker with Podman
Whilst working on Fedora support for Docker, it appeared that in order
to enable it, we need to mess around with selinux settings, simply
because Docker daemon runs as root by default.

Therefore, instead of messing around with security, a recent and more
secure tool called Podman is installed as well.

It is a drop in replacement of Docker when it comes to CLI's, does not
need a daemon unlike Docker, therefore more secure by default.

The installation is also much easier compared to Docker.

Therefore, the role name is changed to acikgozb.podman and the tasks are
updated.