[fedora] Add GCC installation for Neovim
Since most of the plugins need a valid GCC compiler and in some versions
of Fedora this is not provided, a new task is added to make sure of its
existence.

Also, the overall installation of Neovim changed. Since the binary
itself needs lib code, it is moved under $HOME/lib, and a symlink is
created under $HOME/bin to be in sync with the current convention.