[update-nvim] Remove noice.nvim
In order to keep Neovim as default as possible, noice.nvim is removed.
After using noice for several months, I've come to realize that it does
not add much and the default command line is more than enough.
Moreover, it is better to be in sync with default Neovim looks as much
as possible to have similar experience when SSH'ing remote servers.