[gitflow] Add Git config under $XDG_CONFIG_HOME
The current motivation of my revamped gitflow is to use native Git CLI as much
as possible.

Using Fugitive instead of Lazygit was the first step of
this change.

The second step is to understand what Git CLI offers to users in terms
of customization.

Therefore, a Git config file is created to start this adventure.