[gitflow] Exec monthly theme updates
It seems that it's a lot easier to focus with monochromatic themes,
whilst being easy on the eye at the same time.

Therefore, I'm keeping the trend and only switching to another theme for
primary color changes (blue -> pink).
`darkvoid.nvim` is changed with `nvim-noirbuddy`.

The theme update also consists of updating other tools to sync the
theming of the entire dev environment:

- kitty
- tmux
- nvim
- statusline