feat: Update LSP setup
A custom module is implemented for LSP configuration.
This module is mainly used to set up each LSP client in a more dynamic
way.

The main logic is left untouched.

Other than the LSP update, the redundant plugins are removed.
Also, the floating windows coming from LSP and autocomplete are colored
to be in sync with the overall theme.