feat: Change colorscheme to rose pine
The main colorscheme is changed to rose pine. This change includes:

- Changing the $TERM colors (Kitty)
- Changing tmux status bar colors
- Changing $EDITOR colors (Neovim)

There are small adjustments to make for Neovim, because I use a
monochromatic theme with rose pine primary color. I do not use rose
pine theme because it is way too colorful for me.
These changes will be made in later commits.