feat: Introduce `dotfiles_user` and `dotfiles_user_group` facts
In order to correctly assign the ownership of the generated
configuration files, these facts are assigned to the playbook during
pre-tasks step.

Previous version of the `dotfiles` only used $USER via Ansible `lookup`,
causing duplicated calls and increased verbosity in nearly every
single role.

This is removed and moved to pre-tasks section of the playbook.
Therefore, the naming of the task is changed to be more appropriate.