feat: Restructure `acikgozb.net`
`acikgozb.net` is restructured to follow the new structure of the
project.

The configuration files are stored as templates. They are processed and
put under /files directory and they are distributed to the host
system from there.

Currently, this role only supports Arch Linux hosts with specific
dependencies, and this is intentional.

The requirements, dependencies, and the overall responsibility of the
role is explained in its README.