feat: Add playbook level variables to override default role vars
Some of the configuration files are bound to have user specific
values in them.

Even though having a `dotfiles` project is mostly for personal needs,
user specific values are still defined in the playbook level to maintain
them in a single place.

A sample file is added to have a reference under version control,
therefore the actual user specific variable file is added to .gitignore.