feat: Restructure `acikgozb.tmux`
The role `acikgozb.tmux` is restructured to be in sync with the new
architecture.

The configuration file is stored as a template, and the final result is
stored inside the role. It is also copied under `dotfiles/.config/tmux`.

`dotfiles/.config/tmux` is symlinked to `$XDG_CONFIG_HOME` to apply the
changes.

Also, redundant roles (old `acikgozb.tmux`) and directories (`zsh`) are
removed.