feat: Restructure `acikgozb.nvim`
The role `acikgozb.nvim` is restructured to be in sync with the new
architecture.

In order to better define the scope of the role, the name is changed as
`acikgozb.editor`.

The configuration files for Neovim are added under
`dotfiles/.config/nvim`.

Also, default role variables and playbook level variables are
implemented to control the functionality of the role in a better way
compared to the previous version.

The README is updated with great emphasis in role variables.