feat: Restructure `acikgozb.kitty`
The role `acikgozb.kitty` is updated to be in sync with the new
architecture.

First of all, the name is changed to `acikgozb.term` to define the scope
of this role in a better way.

For the first major release, the role is designed for Arch Linux hosts
only, and it installs and configures `alacritty` on a given host:

- `alacritty` is built from the source.
- `alacritty` man pages are also installed.
- templates are created for configuration files and they are put under
  `dotfiles/.config/alacritty` after processing them.
- A soft link for `dotfiles/.config/alacritty` is created under
  `$XDG_CONFIG_HOME/alacritty` to complete the installation and
  configuration.

  The role is tested successfully by executing multiple times.