feat: Create `acikgozb.cloud`
A new role is created to be in sync with the new architecture.
This role combines `acikgozb.aws`, `acikgozb.hashicorp`, and
`acikgozb.podman` and gathers them in a single namespace called
`acikgozb.cloud`.

The role does not contain any custom configurations for given tools,
therefore templates, files, or `dotfiles/.config` are not used at all.

Currently, it is designed for Arch Linux hosts only like other roles.