feat: Add Zellij installation to dotfiles
With this commit, `dotfiles` now supports installing `zellij` as an
alternative to `tmux`.

Since the role `acikgozb.tmux` supports both `tmux` and `zellij`, the
role is renamed to `acikgozb.multiplexer`.

In order to support this change, a new variable is added to `user_vars`.
By default, `zellij` is used as the terminal multiplexer.