feat: Disable text wrap at text width in Helix
The wrapped lines disrupt relative line jumping.

For example, if the target line is 5 lines above, but the current line
is wrapped as 3 lines, then `5k` acts like `2k`.

As a temporary solution, the text wrapping is disabled for text width.
If the text is too long to fit the screen, the line is still wrapped.