feat: helix: Re-enable text width
The text width was turned off due to the fact that it was only applying
soft wrap for Git commits, so I opted out of it and did manual text
formatting.

However, it seems that there is a command called `:reflow` that hard
wraps the selection, which is what is needed for Git commit messages.

Therefore, soft-wrap for text-width is re-enabled to keep the previous
functionality and the keymap `C-r` is added to hard-wrap the selection.

Signed-off-by: acikgozb <acikgozb@proton.me>