feat(git-alias): Add alias for branch cleanup
Three aliases are added to automate cleaning up "stale" branches.
Stale branches in this context means:

- Branches that were already merged to the main trunk, (1)
- Branches that track non-existing remote branches, (2)
- Local branches that were created and eventually abandoned. (3)

`git bcr` is added for (1) and (2).
`git bcl` is added for (3).
`git bca` is added to do (1), (2) and (3) all at once.

Keep in mind that these aliases are designed to be used for branches
that were merged before.

It does not validate branches to see whether they are merged (and safe
to delete) or not.

Signed-off-by: acikgozb <acikgozb@proton.me>