feat(fonts): Simplify font configuration
This commit simplifies the font configuration by adding four new
variables to `user_vars.sample` that are expected to be provided by
the user:

- `terminal_font_family`
- `terminal_font_style`
- `ui_font_family`
- `ui_font_style`

Right now, `terminal_*` variables are used by Alacritty, and `ui_*`
variables are used by `rofi` only, since they are the only programs as
of now that utilizes fonts.

With this change, this commit also changes the font that is used by
Alacritty to `IosevkaSS14`.

Signed-off-by: acikgozb <acikgozb@proton.me>