feat(theme): Allow users to select existing themes
This commit attempts the tackle the theming issues `dotfiles` has,
particularly when it comes to choosing different themes that comes
included with `dotfiles`.

In order to make it a bit easier than before, `user_vars.sample` is
updated to contain several different variables that are responsible from
the overall theming of the terminal environment of the user.

This environment contains:

- The terminal program (Alacritty)
- The terminal multiplexer (tmux or Zellij)
- The $EDITOR (Helix or Neovim)

This commit tries to fix the issue by introducing the `theme` variable
that allows users to select from the pre-existing themes.

Additionally, this commit also exposes the theming that is used for
zellij's status bar and hyprland.

These programs do not have a theme file to work on, and only allow minor
customization through color codes. So providing the `theme` variable has
no effect for zellij's status bar and hyprland.

Signed-off-by: acikgozb <acikgozb@proton.me>