feat(helix): Update config
The config has been re-visited after using Helix for a couple of months.
Most of the changes are straightforward, but there are a couple of
changes that need explanation:

- Automatic signature help is disabled because on small screens the help
popup blocks the code substantially.

- Basic keymaps using the `%{selection}` expansion are added to allow
users to execute commands and use their outputs within Helix (preferably
from scratch buffers) instead of jumping back to the shell and work with
a different set of key bindings.
The keymaps are also added to allow users to use Helix keybinds while
typing commands, which is currently not possible in command mode `:`
(unlike Neovim).

Signed-off-by: acikgozb <acikgozb@proton.me>